requirements.txt
certifi==2025.1.31
charset-normalizer~=3.3.2
filelock==3.18.0
fsspec==2025.3.2
idna~=3.7
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
oauthlib==3.2.2
pillow==11.1.0
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
setuptools~=72.1.0
sympy==1.13.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tweepy==4.15.0
typing_extensions==4.13.1
urllib3~=2.2.2

defusedxml~=0.7.1
ipython~=9.0.2
zlib~=1.2.13
jupyter~=1.1.1
jupyterlab-pygments~=0.3.0
pip~=24.2
pathlib~=1.0.1
configparser~=7.2.0
attrs~=25.3.0
argparse~=1.4.0
distro~=1.9.0
wheel~=0.43.0
pysocks~=1.7.1
openssl~=3.0.16
cryptography~=42.0.5
ipywidgets~=8.1.5
notebook~=7.3.3
pyyaml~=6.0.2
pytz~=2025.2
overrides~=7.7.0
cffi~=1.16.0
regex~=2024.11.6
python~=3.12.4
decorator~=5.2.1
matplotlib~=3.10.1
pydot~=3.0.4
conda~=24.11.3
tqdm~=4.66.4
scipy~=1.15.2
psutil~=7.0.0
packaging~=24.1
click~=8.1.8
zstandard~=0.23.0
libarchive~=3.6.2
async-lru~=2.0.5
pandas~=2.2.3
simplejson~=3.20.1
platformdirs~=3.10.0
pygments~=2.19.1
tokenizers~=0.21.1
safetensors~=0.5.3
scikit-learn~=1.6.1
transformers~=4.51.1
huggingface-hub~=0.30.1
beautifulsoup4~=4.13.3
pydantic~=2.11.2
fastapi~=0.115.12
starlette~=0.46.1
uvicorn~=0.34.0
sentiment~=0.0.1